+++ libcxx/test/std/containers/sequences/vector/vector.cons/default_noexcept.pass.cpp
@@ -30,8 +30,7 @@ struct some_alloc
     void allocate(size_t);
 };
 
-int main(int, char**)
-{
+TEST_CONSTEXPR_CXX20 bool tests() {
     {
         typedef std::vector<MoveOnly> C;
         static_assert(std::is_nothrow_default_constructible<C>::value, "");
@@ -49,5 +48,14 @@ int main(int, char**)
         static_assert(!std::is_nothrow_default_constructible<C>::value, "");
     }
 
-  return 0;
+    return true;
+}
+
+int main(int, char**)
+{
+    tests();
+#if TEST_STD_VER > 17
+    static_assert(tests());
+#endif
+    return 0;
 }