+++ lldb/include/lldb/Core/LoadedModuleInfoList.h
@@ -95,10 +95,10 @@ public:
   protected:
     bool m_has[e_num];
     std::string m_name;
-    lldb::addr_t m_link_map;
-    lldb::addr_t m_base;
-    bool m_base_is_offset;
-    lldb::addr_t m_dynamic;
+    lldb::addr_t m_link_map = LLDB_INVALID_ADDRESS;
+    lldb::addr_t m_base = LLDB_INVALID_ADDRESS;
+    bool m_base_is_offset = false;
+    lldb::addr_t m_dynamic = LLDB_INVALID_ADDRESS;
   };
 
   LoadedModuleInfoList() = default;